Implement use hook in <Article/>
