+++ app/(routes)/articles/[slug]/page.tsx
@@ -1,30 +1,11 @@
-import { cache } from "react";
+import { cache, Suspense } from "react";
 import { client } from "@/sanity/client";
-import { PortableTextItem } from "@customTypes/PortableTextTypes";
 import { Metadata } from "next";
 
-import { Article } from "@components/Article";
+import { Article, ArticleData } from "@components/Article";
 import { BackButton } from "@components/BackButton";
 import { EndOfPageBackButton } from "@components/EndOfPageBackButton";
-
-type Slug = {
-  current: string;
-};
-type Tag = {
-  value: string;
-};
-export type ArticleData = {
-  _createdAt: string;
-  title: string;
-  jpTitle: string;
-  slug: Slug;
-  description: string;
-  jpDescription: string;
-  tags: Tag[];
-  jpTags: Tag[];
-  mainContent: PortableTextItem[];
-  jpMainContent: PortableTextItem[];
-};
+import { LoadingSpinner } from "@components/LoadingSpinner";
 
 const getArticle = cache(async (slug: string) => {
   const article = await client.fetch<ArticleData>(
@@ -58,53 +39,32 @@ type ArticleProps = {
   };
 };
 
-export async function generateMetadata({
-  params,
-}: ArticleProps): Promise<Metadata> {
-  const { slug } = params;
-  const article = await getArticle(slug);
+// TODO: Find a way to re-implement this, without causing a delay in the rendering of the page
+// Currently, this will stop the page from appearing until getArticle has finished
 
-  return {
-    title: article.title,
-    description: article.description,
-  };
-}
+// export async function generateMetadata({
+//   params,
+// }: ArticleProps): Promise<Metadata> {
+//   const { slug } = params;
+//   const article = await getArticle(slug);
 
-export default async function Page({ params }: ArticleProps) {
-  const article = await getArticle(params.slug);
+//   return {
+//     title: article.title,
+//     description: article.description,
+//   };
+// }
 
-  if (!article) {
-    return <div>Article of slug {params.slug} not found</div>;
-  }
-
-  const {
-    _createdAt: createdAt,
-    title,
-    jpTitle,
-    description,
-    jpDescription,
-    tags,
-    jpTags,
-    mainContent,
-    jpMainContent,
-  } = article;
+export default function Page({ params }: ArticleProps) {
+  const articlePromise = getArticle(params.slug);
 
   return (
     <div className="relative">
       <div className="w-full flex justify-center">
         <div className="w-11/12 max-w-1040">
           <BackButton />
-          <Article
-            createdAt={createdAt}
-            title={title}
-            jpTitle={jpTitle}
-            description={description}
-            jpDescription={jpDescription}
-            tags={tags}
-            jpTags={jpTags}
-            mainContent={mainContent}
-            jpMainContent={jpMainContent}
-          />
+          <Suspense fallback={<LoadingSpinner />}>
+            <Article articlePromise={articlePromise} />
+          </Suspense>
           <EndOfPageBackButton />
         </div>
       </div>
